# server/docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "BlancBleu API",
    "version": "1.3.0",
    "description": "Plateforme de gestion de transports sanitaires non urgents (Nice 06).\n\n**Authentification** : la plupart des routes exigent soit le cookie httpOnly `bb_access` (émis par /api/auth/login, prioritaire côté browser), soit le header `Authorization: Bearer <jwt>` (SPA / scripts). Les routes service-to-service (training-data, model/retrain) utilisent `X-Service-Token`.",
    "contact": {
      "name": "BlancBleu — Support",
      "email": "support@blancbleu.fr"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Développement local"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token JWT obtenu via POST /api/auth/login (utilisable en header Authorization)."
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "bb_access",
        "description": "Cookie httpOnly émis par /api/auth/login. Utilisé en priorité par les clients browser."
      },
      "serviceTokenAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Service-Token",
        "description": "Token partagé Node ↔ Python pour les routes service-to-service (training-data, model/retrain, model/status)."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Erreur interne du serveur"
          }
        }
      },
...
      "Prescription": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "numero": {
            "type": "string",
            "example": "PMT-20260524-0001"
          },
          "patientId": {
            "type": "string",
            "nullable": true
          },
          "transportId": {
            "type": "string",
            "nullable": true
          },
          "medecin": {
            "type": "string",
            "example": "Dr. Bernard"
          },
          "dateEmission": {
            "type": "string",
            "format": "date"
          },
          "dateExpiration": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "motif": {
            "type": "string",
            "example": "Dialyse — 3 fois/semaine"
          },
          "fichierUrl": {
            "type": "string",
            "description": "URL du PDF/image scanné"
          },
          "ocrStatus": {
            "type": "string",
            "enum": ["pending", "processing", "done", "error", "skipped"]
          },
          "extractedData": {
            "type": "object",
            "description": "Champs OCR extraits par le service IA"
          },
          "validee": {
            "type": "boolean",
            "default": false
          },
          "validePar": {
            "type": "string",
            "nullable": true,
            "description": "User ObjectId du valideur"
          },
          "valideAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Requête malformée (paramètre invalide, JSON parse error, etc.)",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Token manquant, invalide ou expiré",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Rôle non autorisé pour cette action",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "NotFound": {
        "description": "Ressource introuvable",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    }
  },
  "security": [
    {
      "cookieAuth": []
    },
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/prescriptions/upload": {
      "post": {
        "tags": ["Prescriptions"],
        "summary": "Uploader une Prescription Médicale de Transport (PMT)",
        "description": "Accepte un PDF ou une image (PNG/JPEG/WEBP, max 10 Mo) en `multipart/form-data`,\nchamp `fichier`. Crée une `Prescription` en statut `ocrStatus=pending`\net déclenche un job OCR asynchrone (Tesseract + spaCy local côté\nmicroservice IA) qui peuplera `extractedData`. Le résultat est récupérable\nvia `GET /api/prescriptions/:id/ocr-result`. RBAC : admin / superviseur /\ndispatcher.\n",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["fichier"],
                "properties": {
                  "fichier": {
                    "type": "string",
                    "format": "binary",
                    "description": "PDF/PNG/JPEG/WEBP ≤ 10 MB"
                  },
                  "patientId": {
                    "type": "string",
                    "description": "Si connu, lie la PMT au dossier patient"
                  },
                  "transportId": {
                    "type": "string",
                    "description": "Lien optionnel vers un transport"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "PMT créée (OCR en cours)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Prescription"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "413": {
            "description": "Fichier trop volumineux (> 10 MB)"
          }
        }
      }
    },
    "/api/prescriptions/{id}/validate": {
      "patch": {
        "tags": ["Prescriptions"],
        "summary": "Valider une PMT (IA + humain)",
        "description": "Marque la prescription comme validée. Workflow : l'IA fait l'OCR\npré-validation, puis un staff valide manuellement les champs\nextraits. Trace `validePar` + `valideAt` pour audit CPAM.\n",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "corrections": {
                    "type": "object",
                    "description": "Corrections apportées aux champs OCR (medecin, dateEmission, motif, …)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "PMT validée",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Prescription"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/prescriptions/{id}": {
      "get": {
        "tags": ["Prescriptions"],
        "summary": "Détail d'une prescription",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Prescription trouvée",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Prescription"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  }
}